# commerce/warehouse-api/src/main/resources/warehouse_openapi.yaml
openapi: 3.0.1
info:
  title: Smart Home Technologies API
  version: '1.0'
servers:
  - url: http://warehouse
    description: Generated server url
tags:
  - name: warehouse
    description: API для обеспечения работы склада онлайн магазина
paths:
  /api/v1/warehouse:
    put:
      tags:
        - warehouse
      description: Добавить новый товар на склад.
      operationId: newProductInWarehouse
#      parameters:
#        - name: newProductInWarehouseRequest
#          in: query
#          description: Описание нового товара для обработки складом.
#          schema:
#            $ref: '#/components/schemas/NewProductInWarehouseRequest'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewProductInWarehouseRequest'
        required: true
      responses:
        '200':
          description: OK
        '400':
          description: Ошибка, товар с таким описанием уже зарегистрирован на складе
          content:
            '*/*':
              schema:
                $ref: >-
                  #/components/schemas/SpecifiedProductAlreadyInWarehouseException
  /api/v1/warehouse/return:
    post:
      tags:
        - warehouse
      description: Принять возврат товаров на склад.
      operationId: acceptReturn
      parameters:
        - name: products
          in: query
          description: Отображение идентификатора товара на отобранное количество.
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: object
                  properties:
                    productId:
                      type: string
                    quantity:
                      type: integer
                      format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: integer
                format: int64
              description: Отображение идентификатора товара на отобранное количество.
        required: true
      responses:
        '200':
          description: OK
  /api/v1/warehouse/booking:
    post:
      tags:
        - warehouse
      description: Зарезервировать товары на складе по корзине.
      operationId: bookingProductForShoppingCart
#      parameters:
#        - name: shoppingCart
#          in: query
#          description: Корзина товаров.
#          schema:
#            $ref: '#/components/schemas/ShoppingCartDto'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShoppingCartDto'
        required: true
      responses:
        '200':
          description: Общие сведения по бронированию
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BookedProductsDto'
        '400':
          description: >-
            Ошибка, товар из корзины не находится в требуемом количестве на
            складе
          content:
            '*/*':
              schema:
                $ref: >-
                  #/components/schemas/ProductInShoppingCartLowQuantityInWarehouse
  /api/v1/warehouse/assembly:
    post:
      tags:
        - warehouse
      description: >-
        Собрать ранее зарезервированные товары к заказу для подготовки к
        отправке.
      operationId: assemblyProductForOrderFromShoppingCart
#      parameters:
#        - name: assemblyProductForOrderFromShoppingCartRequest
#          in: query
#          description: Корзина товаров.
#          schema:
#            $ref: >-
#              #/components/schemas/AssemblyProductForOrderFromShoppingCartRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: >-
                #/components/schemas/AssemblyProductForOrderFromShoppingCartRequest
        required: true
      responses:
        '200':
          description: Общие сведения по бронированию
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BookedProductsDto'
        '400':
          description: Ошибка, товара из бронирования нет в БД склада
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/NoSpecifiedProductInWarehouseException'
  /api/v1/warehouse/add:
    post:
      tags:
        - warehouse
      description: Принять товар на склад.
      operationId: addProductToWarehouse
#      parameters:
#        - name: addProductToWarehouseRequest
#          in: query
#          description: Запрос на добавление определенного количества определенного товара.
#          schema:
#            $ref: '#/components/schemas/AddProductToWarehouseRequest'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddProductToWarehouseRequest'
        required: true
      responses:
        '400':
          description: Нет информации о товаре на складе
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/NoSpecifiedProductInWarehouseException'
  /api/v1/warehouse/address:
    get:
      tags:
        - warehouse
      description: Предоставить адрес склада для расчёта доставки.
      operationId: getWarehouseAddress
      responses:
        '200':
          description: Актуальный адрес склада
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AddressDto'
components:
  schemas:
    DimensionDto:
      required:
        - depth
        - height
        - width
      type: object
      properties:
        width:
          minimum: 1
          type: number
          description: Ширина
          format: double
        height:
          minimum: 1
          type: number
          description: Высота
          format: double
          example: 1
        depth:
          minimum: 1
          type: number
          description: Глубина
          format: double
          example: 1
      description: Размеры товара
    NewProductInWarehouseRequest:
      required:
        - dimension
        - productId
        - weight
      type: object
      properties:
        productId:
          type: string
          description: Идентификатор товара в БД
          format: uuid
          example: 53aa35c8-e659-44b2-882f-f6056e443c99
        fragile:
          type: boolean
          description: Признак хрупкости
          example: true
        dimension:
          $ref: '#/components/schemas/DimensionDto'
        weight:
          minimum: 1
          type: number
          description: Вес товара
          format: double
          example: 1
      description: Запрос на добавление нового товара на склад
    SpecifiedProductAlreadyInWarehouseException:
      type: object
      properties:
        cause:
          type: object
          properties:
            stackTrace:
              type: array
              items:
                type: object
                properties:
                  classLoaderName:
                    type: string
                  moduleName:
                    type: string
                  moduleVersion:
                    type: string
                  methodName:
                    type: string
                  fileName:
                    type: string
                  lineNumber:
                    type: integer
                    format: int32
                  className:
                    type: string
                  nativeMethod:
                    type: boolean
            message:
              type: string
            localizedMessage:
              type: string
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        httpStatus:
          type: string
          enum:
            - 100 CONTINUE
            - 101 SWITCHING_PROTOCOLS
            - 102 PROCESSING
            - 103 EARLY_HINTS
            - 103 CHECKPOINT
            - 200 OK
            - 201 CREATED
            - 202 ACCEPTED
            - 203 NON_AUTHORITATIVE_INFORMATION
            - 204 NO_CONTENT
            - 205 RESET_CONTENT
            - 206 PARTIAL_CONTENT
            - 207 MULTI_STATUS
            - 208 ALREADY_REPORTED
            - 226 IM_USED
            - 300 MULTIPLE_CHOICES
            - 301 MOVED_PERMANENTLY
            - 302 FOUND
            - 302 MOVED_TEMPORARILY
            - 303 SEE_OTHER
            - 304 NOT_MODIFIED
            - 305 USE_PROXY
            - 307 TEMPORARY_REDIRECT
            - 308 PERMANENT_REDIRECT
            - 400 BAD_REQUEST
            - 401 UNAUTHORIZED
            - 402 PAYMENT_REQUIRED
            - 403 FORBIDDEN
            - 404 NOT_FOUND
            - 405 METHOD_NOT_ALLOWED
            - 406 NOT_ACCEPTABLE
            - 407 PROXY_AUTHENTICATION_REQUIRED
            - 408 REQUEST_TIMEOUT
            - 409 CONFLICT
            - 410 GONE
            - 411 LENGTH_REQUIRED
            - 412 PRECONDITION_FAILED
            - 413 PAYLOAD_TOO_LARGE
            - 413 REQUEST_ENTITY_TOO_LARGE
            - 414 URI_TOO_LONG
            - 414 REQUEST_URI_TOO_LONG
            - 415 UNSUPPORTED_MEDIA_TYPE
            - 416 REQUESTED_RANGE_NOT_SATISFIABLE
            - 417 EXPECTATION_FAILED
            - 418 I_AM_A_TEAPOT
            - 419 INSUFFICIENT_SPACE_ON_RESOURCE
            - 420 METHOD_FAILURE
            - 421 DESTINATION_LOCKED
            - 422 UNPROCESSABLE_ENTITY
            - 423 LOCKED
            - 424 FAILED_DEPENDENCY
            - 425 TOO_EARLY
            - 426 UPGRADE_REQUIRED
            - 428 PRECONDITION_REQUIRED
            - 429 TOO_MANY_REQUESTS
            - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
            - 451 UNAVAILABLE_FOR_LEGAL_REASONS
            - 500 INTERNAL_SERVER_ERROR
            - 501 NOT_IMPLEMENTED
            - 502 BAD_GATEWAY
            - 503 SERVICE_UNAVAILABLE
            - 504 GATEWAY_TIMEOUT
            - 505 HTTP_VERSION_NOT_SUPPORTED
            - 506 VARIANT_ALSO_NEGOTIATES
            - 507 INSUFFICIENT_STORAGE
            - 508 LOOP_DETECTED
            - 509 BANDWIDTH_LIMIT_EXCEEDED
            - 510 NOT_EXTENDED
            - 511 NETWORK_AUTHENTICATION_REQUIRED
        userMessage:
          type: string
        message:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    ShoppingCartDto:
      required:
        - products
        - shoppingCartId
      type: object
      properties:
        shoppingCartId:
          type: string
          description: Идентификатор корзины в БД
          format: uuid
          example: 53aa35c8-e659-44b2-882f-f6056e443c99
        products:
          type: object
          additionalProperties:
            type: integer
            description: Отображение идентификатора товара на отобранное количество.
            format: int64
          description: Отображение идентификатора товара на отобранное количество.
      description: Корзина товаров в онлайн магазине.
    BookedProductsDto:
      required:
        - deliveryVolume
        - deliveryWeight
        - fragile
      type: object
      properties:
        deliveryWeight:
          type: number
          description: Общий вес доставки
          format: double
        deliveryVolume:
          type: number
          description: Общие объём доставки.
          format: double
        fragile:
          type: boolean
          description: Есть ли хрупкие вещи в доставке.
      description: Общие сведения о зарезервированных товарах по корзине.
    ProductInShoppingCartLowQuantityInWarehouse:
      type: object
      properties:
        cause:
          type: object
          properties:
            stackTrace:
              type: array
              items:
                type: object
                properties:
                  classLoaderName:
                    type: string
                  moduleName:
                    type: string
                  moduleVersion:
                    type: string
                  methodName:
                    type: string
                  fileName:
                    type: string
                  lineNumber:
                    type: integer
                    format: int32
                  className:
                    type: string
                  nativeMethod:
                    type: boolean
            message:
              type: string
            localizedMessage:
              type: string
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        httpStatus:
          type: string
          enum:
            - 100 CONTINUE
            - 101 SWITCHING_PROTOCOLS
            - 102 PROCESSING
            - 103 EARLY_HINTS
            - 103 CHECKPOINT
            - 200 OK
            - 201 CREATED
            - 202 ACCEPTED
            - 203 NON_AUTHORITATIVE_INFORMATION
            - 204 NO_CONTENT
            - 205 RESET_CONTENT
            - 206 PARTIAL_CONTENT
            - 207 MULTI_STATUS
            - 208 ALREADY_REPORTED
            - 226 IM_USED
            - 300 MULTIPLE_CHOICES
            - 301 MOVED_PERMANENTLY
            - 302 FOUND
            - 302 MOVED_TEMPORARILY
            - 303 SEE_OTHER
            - 304 NOT_MODIFIED
            - 305 USE_PROXY
            - 307 TEMPORARY_REDIRECT
            - 308 PERMANENT_REDIRECT
            - 400 BAD_REQUEST
            - 401 UNAUTHORIZED
            - 402 PAYMENT_REQUIRED
            - 403 FORBIDDEN
            - 404 NOT_FOUND
            - 405 METHOD_NOT_ALLOWED
            - 406 NOT_ACCEPTABLE
            - 407 PROXY_AUTHENTICATION_REQUIRED
            - 408 REQUEST_TIMEOUT
            - 409 CONFLICT
            - 410 GONE
            - 411 LENGTH_REQUIRED
            - 412 PRECONDITION_FAILED
            - 413 PAYLOAD_TOO_LARGE
            - 413 REQUEST_ENTITY_TOO_LARGE
            - 414 URI_TOO_LONG
            - 414 REQUEST_URI_TOO_LONG
            - 415 UNSUPPORTED_MEDIA_TYPE
            - 416 REQUESTED_RANGE_NOT_SATISFIABLE
            - 417 EXPECTATION_FAILED
            - 418 I_AM_A_TEAPOT
            - 419 INSUFFICIENT_SPACE_ON_RESOURCE
            - 420 METHOD_FAILURE
            - 421 DESTINATION_LOCKED
            - 422 UNPROCESSABLE_ENTITY
            - 423 LOCKED
            - 424 FAILED_DEPENDENCY
            - 425 TOO_EARLY
            - 426 UPGRADE_REQUIRED
            - 428 PRECONDITION_REQUIRED
            - 429 TOO_MANY_REQUESTS
            - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
            - 451 UNAVAILABLE_FOR_LEGAL_REASONS
            - 500 INTERNAL_SERVER_ERROR
            - 501 NOT_IMPLEMENTED
            - 502 BAD_GATEWAY
            - 503 SERVICE_UNAVAILABLE
            - 504 GATEWAY_TIMEOUT
            - 505 HTTP_VERSION_NOT_SUPPORTED
            - 506 VARIANT_ALSO_NEGOTIATES
            - 507 INSUFFICIENT_STORAGE
            - 508 LOOP_DETECTED
            - 509 BANDWIDTH_LIMIT_EXCEEDED
            - 510 NOT_EXTENDED
            - 511 NETWORK_AUTHENTICATION_REQUIRED
        userMessage:
          type: string
        message:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    AssemblyProductForOrderFromShoppingCartRequest:
      required:
        - orderId
        - shoppingCartId
      type: object
      properties:
        shoppingCartId:
          type: string
          description: Идентификатор корзины в БД
          format: uuid
          example: 53aa35c8-e659-44b2-882f-f6056e443c99
        orderId:
          type: string
          description: Идентификатор заказа в БД
          format: uuid
          example: bbaa35c8-e659-44b2-882f-f6056e443c99
      description: >-
        Запрос на сбор заказа из ранее зарезервированных товаров из корзины к
        заказу.
    NoSpecifiedProductInWarehouseException:
      type: object
      properties:
        cause:
          type: object
          properties:
            stackTrace:
              type: array
              items:
                type: object
                properties:
                  classLoaderName:
                    type: string
                  moduleName:
                    type: string
                  moduleVersion:
                    type: string
                  methodName:
                    type: string
                  fileName:
                    type: string
                  lineNumber:
                    type: integer
                    format: int32
                  className:
                    type: string
                  nativeMethod:
                    type: boolean
            message:
              type: string
            localizedMessage:
              type: string
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        httpStatus:
          type: string
          enum:
            - 100 CONTINUE
            - 101 SWITCHING_PROTOCOLS
            - 102 PROCESSING
            - 103 EARLY_HINTS
            - 103 CHECKPOINT
            - 200 OK
            - 201 CREATED
            - 202 ACCEPTED
            - 203 NON_AUTHORITATIVE_INFORMATION
            - 204 NO_CONTENT
            - 205 RESET_CONTENT
            - 206 PARTIAL_CONTENT
            - 207 MULTI_STATUS
            - 208 ALREADY_REPORTED
            - 226 IM_USED
            - 300 MULTIPLE_CHOICES
            - 301 MOVED_PERMANENTLY
            - 302 FOUND
            - 302 MOVED_TEMPORARILY
            - 303 SEE_OTHER
            - 304 NOT_MODIFIED
            - 305 USE_PROXY
            - 307 TEMPORARY_REDIRECT
            - 308 PERMANENT_REDIRECT
            - 400 BAD_REQUEST
            - 401 UNAUTHORIZED
            - 402 PAYMENT_REQUIRED
            - 403 FORBIDDEN
            - 404 NOT_FOUND
            - 405 METHOD_NOT_ALLOWED
            - 406 NOT_ACCEPTABLE
            - 407 PROXY_AUTHENTICATION_REQUIRED
            - 408 REQUEST_TIMEOUT
            - 409 CONFLICT
            - 410 GONE
            - 411 LENGTH_REQUIRED
            - 412 PRECONDITION_FAILED
            - 413 PAYLOAD_TOO_LARGE
            - 413 REQUEST_ENTITY_TOO_LARGE
            - 414 URI_TOO_LONG
            - 414 REQUEST_URI_TOO_LONG
            - 415 UNSUPPORTED_MEDIA_TYPE
            - 416 REQUESTED_RANGE_NOT_SATISFIABLE
            - 417 EXPECTATION_FAILED
            - 418 I_AM_A_TEAPOT
            - 419 INSUFFICIENT_SPACE_ON_RESOURCE
            - 420 METHOD_FAILURE
            - 421 DESTINATION_LOCKED
            - 422 UNPROCESSABLE_ENTITY
            - 423 LOCKED
            - 424 FAILED_DEPENDENCY
            - 425 TOO_EARLY
            - 426 UPGRADE_REQUIRED
            - 428 PRECONDITION_REQUIRED
            - 429 TOO_MANY_REQUESTS
            - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
            - 451 UNAVAILABLE_FOR_LEGAL_REASONS
            - 500 INTERNAL_SERVER_ERROR
            - 501 NOT_IMPLEMENTED
            - 502 BAD_GATEWAY
            - 503 SERVICE_UNAVAILABLE
            - 504 GATEWAY_TIMEOUT
            - 505 HTTP_VERSION_NOT_SUPPORTED
            - 506 VARIANT_ALSO_NEGOTIATES
            - 507 INSUFFICIENT_STORAGE
            - 508 LOOP_DETECTED
            - 509 BANDWIDTH_LIMIT_EXCEEDED
            - 510 NOT_EXTENDED
            - 511 NETWORK_AUTHENTICATION_REQUIRED
        userMessage:
          type: string
        message:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    AddProductToWarehouseRequest:
      required:
        - quantity
      type: object
      properties:
        productId:
          type: string
          description: Идентификатор товара в БД
          format: uuid
          example: 53aa35c8-e659-44b2-882f-f6056e443c99
        quantity:
          minimum: 1
          type: integer
          description: Количество единиц товара для добавления на склад
          format: int64
      description: Запрос на увеличение единиц товара по его идентификатору
    AddressDto:
      type: object
      properties:
        country:
          type: string
          description: Страна.
        city:
          type: string
          description: Город.
        street:
          type: string
          description: Улица.
        house:
          type: string
          description: Дом.
        flat:
          type: string
          description: Квартира.
      description: Представление адреса в системе.